requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
babel==2.5.0              # via django-phonenumber-field
django-auth-ldap==1.2.13
django-phonenumber-field==1.3.0
django==1.8.17
djangorestframework==3.6.3
phonenumberslite==8.8.0   # via django-phonenumber-field
psycopg2==2.7.1
py==1.4.34                # via pytest
pycrypto==2.6.1
pyldap==2.4.37
pytest==3.1.3
python-magic==0.4.13
pytz==2017.2              # via babel

# The following packages are considered to be unsafe in a requirements file:
# setuptools                # via pyldap, pytest
